EAFP for markets with no specified end date
PredictIt used to use "N/A" as the market end date if the market didn't
specify an end date (see e.g. market 4444).  Now it uses "NA".  Since
we can't rely on PredictIt to provide any particular value to indicate
that a market has no end date, we shouldn't hardcode one in (or a list
of them, for that matter).  Instead, the "easier to ask forgiveness than
permission" approach should be more robust to their changes.